requirements.txt
langchain
langchain-core
langchain-community
crewai
langchain-groq
Pandas
crewai[tools]
flask
gunicorn
PyYAML
pydantic
python-dotenv